--- admin_tools/__init__.py
@@ -1,1 +1,0 @@
-from .topics import topic_exists, create_topic
